Create src
Create a 'for' loop using a range of numbers from 1 to 1000 inclusive. Sum all the numbers that can be divided by both 3 and 5. Print out the numbers that have met the conditions above. Break out the loop once. You have found 5 numbers that have met the conditions above. After breaking out of the loop, print the sum of the numbers that met the conditions above.